Define four key-bindings for common Tup commands
This patch adds a function to execute the Tup program from the shell
along with a macro to bind a key sequence to a Tup command.  The patch
provides the following commands for these keys:

1. C-c C-i: tup init
2. C-c C-u: tup upd
3. C-c C-m: tup monitor
4. C-c C-s: tup stop

There is currently no way to change the bindings except by modifying
the source code itself.