Fix syntax highlighting for comments
Trying to use a regular expression to match comments and apply the
comment face to them breaks as soon as we put any quoted string inside
of the comment line.  Also it is just the wrong approach to take.  It
is more effective and proper to tell Emacs about the format of
comments in tupfiles by modifying the tup-mode syntax table.  This
patch adds entries in the syntax table for the pound character and the
newline so that tup-mode will apply the correct highlighting to
comments no matter what their contents.