Add docstrings for all functions bound to keys
The current definition of tup/make-command-key-binding offers no way
to provide docstrings for the anonymous functions it binds to keys.
This patch changes that macro so that it creates named, non-anonymous
functions to bind to those keys and requires a docstring to associate
with them.  This has two benefits.  First, the name of the function
appears next to the keys in the output of ‘C-h m’.  And second, if the
user clicks on a function in ‘C-h m’, or runs ‘C-h f’ on any of those,
he or she will see a short docstring explaining that function.