Do not bind a key when creating a Tup-command function
This patch is the first step towards collecting all key-bindings in
one place within the source code.  We create a function for each Tup
command via the tup/make-command-key-binding macro.  In order to put
all key-bindings in one place we must first remove that aspect from
the macro.  This patch changes the macro so that it only defines the
Elisp function for running a Tup command.  The patch also renames the
macro to tup/make-command-function to reflect the change in behavior.