Group all key-bindings together in one place
Previously all key-bindings were in different places and that made it
difficult to look at the code and see exactly which keys it claimed.
This patch groups all of the bindings into one variable definition so
that we can not only easily see the key-bindings, but also easily add
new ones without having to hunt through the code for the right place.
The patch also calls ‘use-local-map’ when activating Tup mode to
ensure that the key-bindings take effect.